lowercase package names
