+++ backend/src/main/java/de/neuefische/henny/reisekasse/db/UserDb.java
@@ -1,6 +1,6 @@
-package de.neuefische.henny.reisekasse.Db;
+package de.neuefische.henny.reisekasse.db;
 
-import de.neuefische.henny.reisekasse.Model.User;
+import de.neuefische.henny.reisekasse.model.User;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 public interface UserDb extends PagingAndSortingRepository<User, String> {
