+++ backend/src/main/java/de/neuefische/henny/reisekasse/model/Expenditure.java
@@ -1,5 +1,6 @@
-package de.neuefische.henny.reisekasse.Model;
+package de.neuefische.henny.reisekasse.model;
 
+import de.neuefische.henny.reisekasse.model.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -7,18 +8,19 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.time.Instant;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-@Document(collection = "event")
-public class Event {
+@Document(collection = "expenditures")
+public class Expenditure {
     @Id
     private String id;
-    private String title;
-    private List<EventMember> members;
-    private List<Expenditure> expenditures;
-
+    private List<UserDto> members;
+    private UserDto payer;
+    private Double amount;
+    private Instant timestamp;
 }
