+++ backend/src/test/java/de/neuefische/henny/reisekasse/service/UserServiceTest.java
@@ -1,7 +1,7 @@
-package de.neuefische.henny.reisekasse.Service;
+package de.neuefische.henny.reisekasse.service;
 
-import de.neuefische.henny.reisekasse.Db.UserDb;
-import de.neuefische.henny.reisekasse.Model.User;
+import de.neuefische.henny.reisekasse.db.UserDb;
+import de.neuefische.henny.reisekasse.model.User;
 import org.junit.jupiter.api.Test;
 
 import java.util.Optional;
